requirements.txt
Flask==2.0.1
numpy==1.21.1
pandas==1.3.1
scikit-learn==0.24.2
faiss-cpu==1.7.1
matplotlib==3.4.2
seaborn==0.11.1